KDSingleApplication: Do not exceed maximum socket name length
In cases where there is a long name given (usually the application name) or a long username, `QLocalServer::listen` fails because `sockaddr_un.sun_path` has a char array of 108 on Linux, and 104 on BSD/macOS.
On macOS, the TMPDIR path is long, ie.: `/var/folders/zg/b242mhvd125344h5zrnpkzk80000gn/T/`, so 49 characters just for the temp path alone.
To fix this, use the uid instead of the username if the total length of the socket name is expected to exceed the maximum length.
If it still exceeds the maximum allowed length, ie.: the given name is too long, chop the socket name which cuts off the last part of the name.

I've tested this to work on both Linux and macOS, Windows should be unaffected by these changes.

Fixes https://github.com/strawberrymusicplayer/strawberry/issues/1603